+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 * Add support for `textDocument/prepareRename` request;
 * Upgrade Microsoft.Build + Roslyn libs;
-* update textDocument/documentSymbol handler:
+* Update textDocument/documentSymbol handler:
   - emit hierarchical symbol tree, where client supports it;
-  - emit DocumentSymbol[] instead of SymbolInformation[], the former of which has more info;
+  - emit DocumentSymbol[] instead of SymbolInformation[], the former of which has more info.
 
 ## [0.5.6] - 2022-09-05 / Birštonas
 * Fix an issue where we would fail to generate code action list if one of code action providers would raise an exception; -- log the message to the log instead and continue with the next provider to be more resilent;